# RoamOpenAPI/Sources/RoamOpenAPI/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Roam API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ingest": {
      "post": {
        "summary": "Createingestionjob",
        "operationId": "createIngestionJob_api_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createIngestionJob_api_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Createingestionjob Api Ingest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/{jobId}": {
      "get": {
        "summary": "Getingestionjob",
        "operationId": "getIngestionJob_api_ingest__jobId__get",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jobid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/places/search": {
      "get": {
        "summary": "Searchplaces",
        "operationId": "searchPlaces_api_places_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead",
                  "nullable": true,
                  "title": "Response Searchplaces Api Places Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/places/search-list": {
      "get": {
        "summary": "Searchplaceslist",
        "operationId": "searchPlacesList_api_places_search_list_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 15,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceRead"
                  },
                  "title": "Response Searchplaceslist Api Places Search List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_createIngestionJob_api_ingest_post": {
        "properties": {
          "reelUrl": {
            "type": "string",
            "title": "Reelurl"
          },
          "shareText": {
            "type": "string",
            "nullable": true,
            "title": "Sharetext"
          },
          "reelTitle": {
            "type": "string",
            "nullable": true,
            "title": "Reeltitle"
          },
          "ogDescription": {
            "type": "string",
            "nullable": true,
            "title": "Ogdescription"
          },
          "ogKeywords": {
            "type": "string",
            "nullable": true,
            "title": "Ogkeywords"
          },
          "thumbnail": {
            "type": "string",
            "format": "binary",
            "nullable": true,
            "title": "Thumbnail"
          },
          "frames": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Frames",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "reelUrl"
        ],
        "title": "Body_createIngestionJob_api_ingest_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "reelUrl": {
            "type": "string",
            "title": "Reelurl"
          },
          "shareText": {
            "type": "string",
            "nullable": true,
            "title": "Sharetext"
          },
          "reelTitle": {
            "type": "string",
            "nullable": true,
            "title": "Reeltitle"
          },
          "ogDescription": {
            "type": "string",
            "nullable": true,
            "title": "Ogdescription"
          },
          "ogKeywords": {
            "type": "string",
            "nullable": true,
            "title": "Ogkeywords"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "title": "Error"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "pipelineResult": {
            "$ref": "#/components/schemas/PipelineResultRead",
            "nullable": true
          },
          "pipelineResults": {
            "items": {
              "$ref": "#/components/schemas/PipelineResultRead"
            },
            "type": "array",
            "title": "Pipelineresults"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "reelUrl",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "IngestJobResponse",
        "description": "GET /ingest/{jobId}: one reel job can produce multiple ideas (one pipeline_result per candidate)."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "processing",
          "done",
          "failed"
        ],
        "title": "JobStatus"
      },
...
      "PipelineResultRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ideaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Ideaid"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Jobid"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "refinedTitle": {
            "type": "string",
            "nullable": true,
            "title": "Refinedtitle"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "title": "Category"
          },
          "estimatedMinutes": {
            "type": "integer",
            "nullable": true,
            "title": "Estimatedminutes"
          },
          "modelName": {
            "type": "string",
            "nullable": true,
            "title": "Modelname"
          },
          "promptVersion": {
            "type": "string",
            "nullable": true,
            "title": "Promptversion"
          },
          "rawOutput": {
            "additionalProperties": true,
            "type": "object",
            "nullable": true,
            "title": "Rawoutput"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "placeSuggestions": {
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestionRead"
            },
            "type": "array",
            "title": "Placesuggestions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "createdAt"
        ],
        "title": "PipelineResultRead"
      },
      "PlaceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "googlePlaceId": {
            "type": "string",
            "nullable": true,
            "title": "Googleplaceid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "title": "City"
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "title": "Longitude"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "title": "Category"
          },
          "openingHours": {
            "items": {},
            "type": "array",
            "nullable": true,
            "title": "Openinghours"
          },
          "placeTypes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "nullable": true,
            "title": "Placetypes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "estimatedMinutes": {
            "type": "integer",
            "nullable": true,
            "title": "Estimatedminutes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "createdAt"
        ],
        "title": "PlaceRead"
      },
      "PlaceSuggestionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "resultId": {
            "type": "string",
            "format": "uuid",
            "title": "Resultid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "rawName": {
            "type": "string",
            "nullable": true,
            "title": "Rawname"
          },
          "confidence": {
            "type": "number",
            "nullable": true,
            "title": "Confidence"
          },
          "isSelected": {
            "type": "boolean",
            "title": "Isselected"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resultId",
          "isSelected",
          "createdAt"
        ],
        "title": "PlaceSuggestionRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}